Keep old track IDs when moving files within the library
Our previous FileHooks implementation was such that moving a scanned audio
file from on folder within the library to another folder within the library
caused the track object matching the said file to be deleted and the
recreated. Hence, the track got a new ID and it got removed from any
playlists where it belonged to.

Now we handle file moving separately and differently than a file deletion
followed by file recreation, and the track gets updated "in place", without
changing the ID or removing it from existing playlists.

If, on the other hand, an entire folder full of scanned audio files gets
moved, then we do one of two things:
1) In case there's 30 or less audio files contained, we handle each file
separately the same way as if those 30 files were moved individually
2) In case there are more than 30 audio files contained, we remove them all
from the library and the user needs to rescan the files.

The second case is suboptimal but we can't handle any number of files
one-by-one during the moving because that could slow down the moving
operation too much. In the best case this would annoy the user and in the
worst case it would cause a time-out and break the entire operation.

This is a partial fix for https://github.com/owncloud/music/issues/1173.