Refactoring: Add type declarations to class members
PHP 7.4 added support for type declarations also on the class members. As we
have required PHP 7.4+ since Music 2.0.0, we are now free to use them.

Type declarations were added to almost all class members. As a notable
exception, the members matching the DB columns in the Entity classes were
excluded. This was because all the typed members should be initialized in the
construction but these members are injected from the DB through some framework
magic.