Refactoring: Access SubsonicController::$userId with null check
With the type declarations now in place, Scrutinizer realized that
SubonicController::$userId may technically be null although that is not a
valid userId for most functions. In practice, our check in
SubsonicController::handleRequest prevents using the null value in places
where it's not acceptable but Scrutinizer doesn't know this.

Hence, we now access the $userId member almost everywhere through a private
method SubsonicController::user() which contains a null check.