Refactoring: Add helper for getting the current user for AmpacheController
Similarly as previously for SubsonicController, we need this to assert to
Scrutinizer that we actually have a valid session whenever we want to get
the current user ID. Also, this makes the code a tiny bit more concise and
less repetitive.

Apart from getting the current user ID, there are also a few other cases
where $this->session needs to be accessed. These are now asserted separately.
As with the use ID, the middleware should take care that these functions will
never get called without a valid session.