Refactoring: Use PHP arrow functions where it makes sense
The arrow functions were introduced in PHP 7.4 which is our oldest supported
version as of Music 2.0.0.

In this commit, all the invocations of \array_map and \array_filter were
checked and the arrow function was applied to those where it was a good fit.
There might remain some other callback cases which could still benefit from
the use of arrow function.