Extract localizable strings also from back-end and non-angular front-end
For this, we use the l10n/l10n.pl script which was slightly modified so that
it combines its results to previous music.pot (from angular-gettext) instead
of overwriting it. It was also modified to roam only the non-agular source
directories. This new extraction now happens as part of `make l10n-extract`.

In addition to perl and its Locale::PO module, the l10n.pl requires the
xgettext tool. Some pointers about this and the general use of the
translation scripts was added to README.md.

As we now found the strings from the actual sources, many of the entires in
the fake-template.php could be removed. The few remaining entries are such
that the string is defined in an angular template as part of a more complex
statement.