Make l10n.pl handle .js/.ts files as JavaScript instead of faking it's Python
To have this supported, the xgettext has to be of sufficiently new version.
At least the version 0.22.5 has the support while the version 0.18.1.1 does
not.

With this, we could extract one more string "Select mode" from the dashboard
widget. This previously failed with the Python mode presumably because there
was a # character on the same row before translation and this annotates a
comment in Python. It's a small wonder that the Python mode worked otherwise
so well here.