Skip background cleanup while any scan job is running
With MariaDB, it sometimes happened that executing the background cleanup job
while scanning the music library caused the said scan job to fail with the
message "SQLSTATE[40001]: Serialization failure: 1213 Deadlock found when
trying to get lock; try restarting transaction". To prevent this, the cleanup
job now skips touching the library DB tables if there are any ongoing scan
jobs.

It was first attempted to prevent only touching the DB rows of the specific
user doing the scan but this was not enough to solve the issue. Presumably,
it's now some table-level locking which causes this issue.

The fix is not completely thread-safe in the sense that if multiple scan
jobs are run simultaneously for the same user, then the cleanup job may still
get to access the tables. But this is quite pathological case, so probably
not a real-life problem. And even if this happens, then one of the scan jobs
just fails and the other one probably runs normally.

refs https://github.com/owncloud/music/issues/918