Ampache API: Enable cross-origin calls
When a web app Ampache client such as Ample connects to the server from
another domain, the server must set certain headers to allow such access.
Also, the web app may do a "pre-flight" first by using the HTTP OPTIONS call
on the API endpoint, and the server must support this.

The Nextcloud server already contained support for such cases but a few steps
were needed to activate it on our Ampache endpoints:

- The endpoint functions need to be annotated with @CORS. This is a sign for
  the CORSMiddleware in the server core to add the header
  'Access-Control-Allow-Origin'

- AmpacheController has to be inherited from ApiController and the allowed
  CORS headers need to be setup in the constructor. In addition to the
  default set of allowed headers, Ample requires 'X-Requested-With' as
  documented in
https://github.com/mitchray/ample/wiki/Connecting-to-a-remote-Ampache-server-(CORS)

- The base class ApiController provides the pre-fligth support with the
  function `preflightedCors` which needs to be mapped in routes.php.
  AmpacheMiddleware must not attempt to do any authentication checking on the
  pre-flight calls

The Nextcloud documentation for this can be found from
https://docs.nextcloud.com/server/30/developer_manual/digging_deeper/rest_apis.html

refs https://github.com/owncloud/music/issues/1181