Support range requests when relaying external streams
The relevant HTTP headers are relayed between the client and the server
providing the stream. Hence, the range requests are supported if and only if
the stream provider supports them.

Range requests are probably never supported on live radio streams but they
are vital when streaming podcasts where the episodes may be rather long and
jumping tens of minutes forward may be desirable.