Make radio and podcast relaying configurable
There are now two new keys available for config.php:
'music.relay_radio_stream'
'music.relay_podcast_stream'
The default value is true for both of these.

Also, the default value of the key 'music.allowed_radio_src' is now an empty
array, meaning that the cloud admin has to specifically opt in to any sources
he wishes to allow (in case the relaying is disabled). The admin still may
define the allowed source as ['http://*:*', 'https://*:*'] to allow
everything.

The admin settings descriptions in the Settings view have not been updated
yet.