Improve readability of the dashboard widget play timer
The song duration and current play time were difficult to read on many color
schemes when the progress bar was behind the text. Adding a hazy shadow on
the text improves the situation a bit. The color of the shadow is the same
light-gray as the background color of the progress bar.