Security hardening: Add security tokens to HLS streaming endpoints
URL arguments of the the API endpoints `/api/radio/hls/manifest` and
`/api/radio/hls/segment` are now "signed" by the server by adding a security
token which is hashed from the URL, a private secret known only by the server,
and a random salt. When these endpoints are then called, the server checks
the authenticity of the URL against the security token.

With this, we can ensure that a malicious or compromised client cannot trick
the server to make HTTP calls to just any URL, hence effectively bypassing
the Content-Security-Policy set on the front-end.