Relay the podcasts in the Ampache API only for the internal client
Eventually, there could be one or more additional setting options to enable
the relaying of podcasts and radio streams for the external Ampache and
Subsonic clients. But as that is not yet available, and not a priority right
now, it was a bit weird that podcasts got relayed but radio streams did not.
And this applied to Ampache API but not to Subsonic API.