Move the admin settings documentation to the Wiki
This is very technical and having it as translatable content of the app
itself carries a significant risk that something is lost in translation.
Also, it was a burden for any potential translators wishing to create a
complete translation of the app. And now the wordings can be fine-tuned
(if necessary) without worrying about this breaking the existing translations.