Update radio "Getting started" and move troubleshooting to wiki
The "Getting started" guide had some not quite up-to-date information about
the use of the Community Radio Browser which was now replaced with a bit more
generic instructions. Also, the guide now mentions also the possibility to
add the stream URL manually.

The radio troubleshooting is no longer part of the "Getting started" pane;
instead the pane offers a link to the topic in the wiki. The previous
troubleshooting content are now irrelevant for the majority of the users as
the streams are relayed via the cloud by default.