Help VSCode to not show so many false positive errors
VSCode doesn't understand how we use class_alias to create the base class
CompatibleMapper. Hence, it reported an error for each line where any of the
methods of this base class were called. To work around this, we now annotate
the used methods from CompatibleMapper on the BaseMapper using PHPDoc.

This is a hack, especially because some of the annotated functions are
actually protected but PHPDoc only has syntax to declare public methods. The
downside of this solution is that VSCode now thinks that these protected
methods could be called from outside and offers such option on the code
completion.