Subsonic: `getMusicDirectory` to sort by file name when browsing by folders
Previously, the results were sorted by song title which was suboptimal for
example in case the folder contained songs of one album and file names were
prefixed with the track numbers. Or if the library had flat structure with
everything in one folder with file naming like "Artist Name - Song Name" or
similar.

Also, when browsing by meta data, the file name is now used as the third
sorting criterion after the disk number and track number.

refs https://github.com/owncloud/music/issues/1182