Subsonic: Include property `path` to all song responses
The downside of this is that injecting the path info may take a significant
amount of time in the extreme case where the user has a huge library (with
thousands of folders) in an externally mounted storage, and the Subsonic
client fetches a large amount of songs (e.g. all of them using an empty
`search2` query). The reason for this is that none of our optimizations for
folder resolving are used in the case of externally mounted storage. On the
other hand, also the Folders view on the web UI is extremely slow to load in
this case. This is something which should be worked on.