Optimize folder tree building also with external mounts under the root
Previously, we had optimized resolving for the folder hierarchies only for
those parts of the library which belonged to the same storage as the library
root. Hence, loading the Folders view (or large Subsonic song sets) could get
insanely slow if the home folder was selected as library root, and there was
a huge externally mounted audio folder under it. For example on our test
setup with 12000 folders on external mount, loading the Folders view could
take more than one minute. If the said mount was selected as the library
root, then there was no performance problem.

Now the optimization logic has been rewritten so that it gets utilized also
in those cases where there are external mounts under the library root. This
makes the previously mentioned 12000 folder tree load in a few seconds. Also,
the impact of adding the 'path' to huge Subsonic song responses is now
negligible.  The only query now made directly using the API is to get all the
direct children of the library root. Everything else is fetched with custom
SQL queries directly from oc_filecache.