Don't use rowCount() on SQL SELECT result as it doesn't work on SQLite
As stated on https://www.php.net/manual/en/pdostatement.rowcount.php,
PDOStatement::rowCount is guaranteed to produce meaningful results only for
DELETE, UPDATE, and INSERT statements. Most importantly for use, it always
returns 0 for SELECT queries on SQLite. The way we used it, happened to work
correctly on MySQL and PostgreSQL.

This caused several bugs when running the app with SQLite:

- Ampache actions `playlist_edit`, `playlist_add`, and `playlist_add_song`
  didn't work

- Image file uploaded to a folder containing an album without cover didn't
  get picked up immediately and used as cover. It only happened once the
  backgorund job ran.

- Background cleanup prevention during scanning didn't work (1b0ec6b6)