Fix Dashboard widget not reacting to song click after stop
After stopping the playback through the media session integration (e.g. a
stop key on the keyboard), the same song didn't start playing again by
clicking it. Instead, the user had to play some other song first. This was
because the state of the queue was not cleared on stop, and this made the
widget believe it should just toggle play/pause of the current song (although
the PlayerWrapper no longer had any source url).