Fix radio streaming from URL containing no path and no trailing '/'
The radio streaming and metadata fetching didn't work for such stream URL
where there was only the domain and possibly a port number but no any kind of
path after that. If there was even the "empty path" denoted with a trailing
slash ('/'), then things worked fine.

As a working example, https://runder.org:8000 is a stream URL where this bug
used to show.