Fix header parsing in HttpUtil::loadFromUrl
In case there was a redirection in the URL loaded, the status_code and
message returned came from the first redirection response and not from the
last response after the redirection(s). Potentially, also the content_type
could come from an intermediate result although likely the redirection
responses shouldn't define this header.

This may have caused other bugs too, but at least this broke the radio
playback in cases where the given URL redirected to another URL which was a
playlist file pointing to the actual audio stream. For example,
http://sverigesradio.se/topsy/direkt/226-hi-aac.pls is such URL.