Basic support for Nextcloud 31
NC 31 has removed the method Server::getLogger() as well as all methods from
the ILogger interface. The new alternative for this has existed since NC 20
but it's not there in ownCloud. Hence, we need to be prepared to two
different environments.

With this change, at least the basic functionality of the Music app works.
Everything has not been tested yet.

refs https://github.com/owncloud/music/issues/1198