Fix embedded player on publicly shared folders on NC31
In Nextcloud 31, also the public link-shared folders use the new Vue UI
instead of the legacy Files UI. Hence, the "NC28+" logic in the embedded Music
player now needs to be able to handle also the public shares.

To easily stream the file data from public shares, both when playing an audio
file and when playing a playlist, a new endpoint
`/api/share/{token}/{fileId}/download` was added. The same endpoints provided
by the Music app back-end are now used regardless of the host cloud version.

refs https://github.com/owncloud/music/issues/1198