Fix narrow screen navigation pane layout on NC30
When the browser window is narrow and the navigation pane got collapsed, it
was left peeking on the left border of the window on Nextcloud 30. This was
related to the fact that NC30 shows the borders around the content also in
the narrow screen "mobile" layout while NC25-29 removes the borders in this
mode. Interestingly, NC31 removes the borders of mobile layout once again so
NC30 is the only outlier.

As an additional glitch, the bottom left corner of the content was shown
without the normal rounding when the collapsed navigation pane was expanded.

The major change made here is to use "contain: content" on the #app container.
With this, we don't anymore need to manually handle the borders and rounded
corners as those will be cropped automatically. This change required making
changes to how the layout for navigation pane, controls pane, and the sidebar
was defined. What made this somewhat difficult was that this rule cannot be
used on old Nextcloud versions (14 to 24) because of how they use the
window.document as the main scroll container, and we still support NC versions
20+. Hence, we needed to add some exceptios for these versions.