Ampache: Parse and include lyrics in the response of the action `song`
We don't store the lyrics in the DB so we can't return them in each
song-type response like intended. However, when fetching just one song, it's
not really an issue to parse that one file at that time and include the
lyrics. This seems to work great at least on Ample v3.1.2 as it always
calls the `song` action when the playing song changes and when an individual
song is opened for details.