Manually handle redirections when relaying external streams
At least the podcast "No such thing as a fish"
(https://audioboom.com/channels/2399216.rss) failed to play when the relayed
streaming mode was used (as is the default). One reason was that the stream
URLs of that channel had a whopping 7 redirections before landing to the
actual stream, and we allowed maximum of 5 redirections. However, just
increasing the allowed number of redirections to 20 didn't solve the problem,
as then the stream was returning 403 "Forbidden" with a message stating that
the URL argument Key-Pair-Id would be missing (which was part of the
target URL of the redirection).

The problem got solved when the automatic redirections were disallowed in the
PHP stream context. Instead, the redirections are now handled by our own
logic within RelayStreamResponse. The exact reason for the automatic
redirection not working here was left as a mystery.

refs https://github.com/owncloud/music/issues/1209