Subsonic: Fix `getPlaylist` failing on invalid song references
The Subsonic method `getPlaylist` failed ungracefully with HTTP 500 and
logged an error if the field `trackIds` contained any such ID which did not
match an existing track belonging to the user. Normally there should never be
such track references but some rare special/error cases could induce them.

The errors logged were quite difference between PostgreSQL and MariaDB as on
the former the flow failed a bit earlier than in the latter.

refs https://github.com/owncloud/music/issues/1128