Dashboard widget: Fix album art on radio station without given name
An error happened when trying to create the placeholder album art for a radio
station without given name. This was because a null was returned by a
function which was always supposed to return a string. As a result, the
loading indicator was shown indefinitely instead of the album art.