REST API: Make endpoint `playlists/{id}/add` Shiva-compliant
The endpoint `playlists/{id}/add` can now be used according the Shiva
specification, allowing the caller to define the insertion index. However,
we define the `index` argument as optional and omitting it appends to the end
of the list as before.

Also, the name of the argument for the track IDs is now `track` instead of
`trackIds` to follow the spec. But unlike in the Shiva spec, we still support
giving there more than one ID as a comma-separated list.