REST API: Make endpoint `playlists/{id}/remove` Shiva-compliant
To comply with the Shiva specification, the argument name is now `index`
instead of `indices`. The specification defines support only for removing one
track at a time but our implementation still allows passing a comma-separated
list instead. This can be considered an extension to the documented API, as a
single index is just a 1-item special case of a list.