Fix podcast episode titles overlapping alphabet navigation on narrow screen
It's actually a bit of mystery, why changing the "display" property of the
container element helped here. It didn't remove the problem quite completely
but it reduced it a lot.

Also, it's a bit weird that this problem was clearly visible in the Podcasts
view but not in the Albums view while both use basically the same CSS. The
current assumption is that the reason is podcasts routinely having longer
episode number and also longer titles.