Fix Subsonic API not working on PHP versions 7.4 and 8.0
The database integration seems to work a bit differently on PHP 8.0 and older
compared to 8.1 and newer: When calling
\Doctrine\DBAL\Driver\Statement::fetch, the newer PHP versions seem to cast
the integer-typed DB column values automatically to ints while older versions
return them as strings. As we later assume the values to be ints, an explicit
cast was added.

The type mismatch prevented entirely the use of the Subsonic API on PHP 8.0
and older. This bug has got introduced by the commit dce4bea8 first released
in Music v2.1.3.

refs https://github.com/owncloud/music/issues/1218