Fix disinformation from the prompt on export file name conflict
The pop-up used to say that the exporting could be canceled by closing the
pop-up (without making any selection). However, this doesn't actually work on
any of the supported Nextcloud versions where closing the dialog does the
same thing as the "No" button. On ownCloud, it did work as intended.

The text was also modified a bit to be more reusable, in case we need to later
save also other things than the exported playlists.