Allow manual entry of file name when exporting playlist or radio stations
When exporting a playlist or the internet radio stations to a file from the
web UI, the UI now queries the file name to use from the user. The default
name is the playlist name or "Internet radio" with the extension .m3u8 and
this is prefilled in the text field.

In case of name conflict, and user selecting "No" to the overwrite query,
there is now a new name query with the suggested name (including an
appropriate number postfix) prefilled.