Subsonic: Use error code 0 instead of 70  when unsupported endpoint requested
When a Subsonic client requests an API endpoint which is unsupported or
misspelled, the server now responds code 0 "A generic error" instead of the
previously used 70 "The requested data was not found". The latter is now used
only for the cases where the request is otherwise fine and serviceable but
the entity specified by the parameters is not found.