Util::truncate now operates on the number of bytes instead of characters
When we use the function Util::truncate, the purpose is usually to make the
data fit in the column of a DB table. But when using a multi-byte encoding
like utf8, it's not necessarily enough to cut the string to the number of
characters matching the column width of the DB table. This may depend on the
DBMS and its settings, though.