Fix the filename truncation logic on playlist export
There was a problem in how truncating the filename to maximum length
interacted with the conflict resolution. Example scenario if we imagine
that the maximum length for the filename would be 15 characters and we
reserve 5 characters for the number postfix " (xx)":
- User gives filename "1234567890123"
- It gets truncated to "123456789…" and saved
- On next export, user gives again name "1234567890123"
- It gets truncated to "123456789…" but there is a conflict => the suggested
  name is "123456789… (2)"
- User accepts the suggested name but it gets again truncated to "123456789…"
  which is again a conflict

To resolve this, we no longer reserve the extra 5 bytes for the number
postfix if there is one already. The server core logic will anyway use and
increment the existing postfix instead of appending a new one.