Fix alphabet navigation having some offset on first click in list views
In long lists of items using the "two-line layout with album art" (like "All
tracks"), the first click on the alphabet navigation target didn't scroll to
quite correct spot. Second or third click on the same target fixed this. The
reason was that the placeholders for the off-screen parts of the list had
their height calculated incorrectly. The directive aways used the height of
the one-line layout, regardless of the layout actually used by the instance.

Also the height value used for the one-line layout items was changed from
31.4833px to 31px to match the present-day height of the actual list items.
It's not known, when has this actually changed.