Redesign the unique constraint handling in Mappers
The way we previously handled the unique constraints on DB updates caused a
problem in case of rescan and any of the scanned tracks was already in the DB
but with a broken `album_id` or `artist_id` reference. In such case, the
rescan was terminated with the error "Did expect one result but found none".
This happened when the function BaseMapper::insertOrUpdate tried to fetch the
existing track with the same `file_id` but the broken reference caused that
to fail.

In the new logic, we don't try to reconstruct the existing entity but fetch
only its ID and creation timestamp. In addition to avoiding the problem
described above, this saves a significant amount of time. This is most
visible when Redis is used for transactional file locking, avoiding other
unnecessary slow downs. In such setup, the scan time was seen to improve by
~50 % on MariaDB and by 20-30% on PostgreSQL and SQLite.

refs https://github.com/owncloud/music/issues/1228