Refactoring on SubsonicController
Instead of calling SubsonicController::subsonicResponse in each and every
@SubsonicAPI method, those methods now mostly return an array. The dispatcher
function SubsonicController::handleRequest takes care of wrapping the array
as Response if necessary. An individual API method still may return also a
Response object; this is necessary on binary endpoints and in those few cases
requiring a non-default value for the parameter $useAttributes of
SubsonicController::subsonicResponse.

Same kind of architecture has been used in AmpacheController for quite some
time, already.