Slight refactoring on XmlResponse
Call the private function addChildElement so that the first argument is
always an instance of DOMElement. Previously, it could be also DOMDocument.
Now we need to handle the root element as a special case in
XmlResponse::render but we were already doing that in the special case of
empty content.