Refactoring: Declare own abstract class TimedJob instead of using class alias
For hiding the OC/NC difference in base class namespaces, we now declare our
own abstract base class instead of doing the same with a class alias. The
alias was problematic at least on the VSCode IDE which couldn't understand
it. Also PHPStan required some extra configuration to survive with the alias.

Our own base class has the same name TimedJob as the platform class but it
resides in our namespace OCA\Music\AppFramework\BackgroundJob.