Add some type hints to BaseMapper for Scrutinizer and VSCode
Added type hints indicating that the inherited functions are supposed to
operate with instances of OCA\Music\Db\Entity instead of
OCP\AppFramework\Db\Entity. We already had the template hints for PHPStan but
Scrutinizer and VSCode don't understand those.

On the other hand, eclipsing the $entityClass string member from Mapper
didn't seem to be anymore necessary for PHPStan.