Scanner: Slightly refactor handling of int-type track properties
Ensure that all the numeric values get interpreted as integers and never as
floats. On 32-bit systems, values greater than SINT32_MAX will get
implicitly converted to floats upon extraction. However, we anyway limit the
values to SINT32_MAX before storing, because of the limitations of
PostgreSQL. Hence, we can safely cast all the values to integers. With this,
we don't need to consider multiple possible types after the values have been
normalized.