Remove unnecessary use of PHP reference variables and parameters
When dealing with object type variables, it's rarely necessary to use
reference parameters. This is because an object variable in PHP is actually
kind of pointer, and the same underlying data may be accessed even when the
variable itself is passed by value. The gotcha here is that arrays in PHP are
not objects and they act like they were primitive data types on value passing
(although under the hood, the runtime probably uses copy-on-write to avoid
unnecessary data copying).