Scanner: Don't print debug info on disregarded events
Scanner used to emit some debug prints whenever the hook functions triggered,
e.g. for each and every file creation/modification/deletion. Now it's more
conservative, printing something only when the event is somehow related to
the music library. That is, if the event can be quickly discarded because it
happened outside of the music library or is of unrelated file type, then
nothing is logged.