Subsonic: Fix response of `getAlbumInfo(2)` when no info found
In case there was a request `getAlbumInfo` or `getAlbumInfo2` and no info
was available, the value of the property `albumInfo` in the JSON format
result was inappropriately an empty array instead of the more logical empty
object. In the XML format, the element `albumInfo` was absent instead of
being just empty.

refs https://github.com/owncloud/music/issues/1233