Slightly refactor Util::buildUrl
Try to make Scrutinizer not to frown upon it so badly. This actually cheats
the logic of Scrutinizer a bit: it seems to consider the ternary operator ?:
as branching but not the null-coalescing operator ??, although they are
practically very similar. Having "3 less branch points" will make Scrutinizer
believe that there are 256 instead of 2048 paths through this function.