Avoid calling IAppContainer::getServer on Nextcloud
This method has been deprecated since NC20. Furthermore, the method
getContentSecurityPolicyManager of IServerContainer (returned by getServer)
is about to be removed in NC32.

Also our replacement IAppContainer::query is deprecated since NC20 but that
one we use in many other places, too. Replacing it is a task for another
commit.