Provide wrapper for IAppContainer::get/query in our Application class
IAppContainer::query is deprecated since NC20+ while its replacement
IAppContainer::get doesn't exist on ownCloud. There's now a wrapper function
Application::get which always calls the more appropriate of the two. The
direct calls to IAppContainer::query and IAppContainer::get can still be used
in such contexts where we know that the cloud is ownCloud or that it is
Nextcloud.