Refactoring: Psr\Log\LoggerInterface-like interface for Logger
Our Logger class now has a dedicated logging function for each severity
level, the same way as Psr\Log\LoggerInterface and OCP\ILogger. This makes
the code a bit more concise at the point of the call. Unlike those mentioned
platform interfaces, we no longer provide the `log` function with the severity
level as a parameter since we don't have any use for that. Also, our
functions don't have the optional $context parameter as there's currently no
use for that, either.

This will later make it easy to switch to using the Psr\Log\LoggerInterface
directly if we one day drop the support for ownCloud.