Avoid "dirty table read" in AmpacheUserMapper::addUserKey
Writing to the DB and then immediately reading back the written row is a bad
practice since it may fail on very large installations with a distributed DB
server. In this particular case, we were interested only about the ID of the
added row and the DB read could be avoided altogether by using the function
IDBConnection::lastInsertId.