Optimize evaluating "dirty" music file IDs
The function trackBusinessLayer::findAllDirty took significant amount of time
on huge music libraries because it internally fetched and constructed full
Track objects for all tracks of the library. It was replaced with
TrackBusinessLayer::findDirtyFileIds which fetches only the file IDs from the
database. All the filtering happens within the DBMS. The difficulty here was
that we need to compare datetime strings to Unix timestamps and the
conversion between the two is different on each supported DB back-end.