Fix music controls pane overlapping scrollbar with Firefox on Windows 11
By default, the Firefox browser uses an "overlay" scrollbar when running on
Windows 11. The browser doesn't reserve any screen space for this kind of
scrollbar but overlays it on the content. This scrollbar used to get hidden
behind the music controls pane, both in the Music app proper and in the Files
app, since the scrollbar width evaluates as 0 and no space is reserved for it.

Getting the actual width of the overlay scrollbar seems to be pretty much
impossible. Even hard-coding it wouldn't really work since it's also
impossible to get the zoom level of the browser and this affects the width of
the scrollbar in terms of "reference pixels".

However, it was found out that the overlay scrollbar gets drawn behind the
pane only if the pane has a non-default z-index set. Hence, this could be
fixed by refactoring the code so that we don't need to use the z-indices.

In the Files player, we couldn't get rid of the z-index on ownCloud and older
Nextcloud versions because components of the files list use non-default
z-indices there. On the other hand, this problem was never visible in Files on
older NC-versions while it was on NC28+ and on all ownCloud versions. Hence,
we now remove the z-index there only on NC28+ which means that the problem
remains only on ownCloud.

In addition, the Music app proper had such a problem that the overlay
scrollbar couldn't be dragged on ownCloud because it was behind the alphabet
navigation pane. Also this got fixed by removing the z-index. For some
reason, this problem didn't show up on NC.