Fix Ampache API key creation showing an error on some environments
Although IDBConnection::lastInsertId should return int according to its
phpdoc, it seems to actually return a string on some environments. This was
reproduced at least on OC10.6, OC10.14, and NC20, all using sqlite3. This
doesn't seem to happen on the recent NC versions. This may be related to the
fact that https://www.php.net/manual/en/pdo.lastinsertid.php actually returns
a string.

This problem was introduced in 3bfbb7ec which has not yet been in any released
app version.