Use the shared volume control widget also in Music app proper
There's no need to define the same logic multiple times. To embed the shared
widget to AngularJS, it was just wrapped in a thin directive. The
functionality is otherwise unchanged but the current volume is no longer
shown in the tooltip; the plan is to show it in place of the volume icon
during the volume adjustments.