Add close button to the "files to scan" banner
With the button, the "New music available" or "Some files may have changed"
banner may be hidden without starting the scan. The banner is shown again
once the page is reloaded. The button is not shown when the library is
completely empty and the "New music available" box is shown in the middle of
the window.

refs https://github.com/owncloud/music/issues/1251