Fix files remaining "dirty" even after rescan when folder moved
The root cause of this was a mismatch between how the platform Entity and
Mapper classes track changed fields and what the Music app assumed and
required. This in combination with changing the `Track::dirty` default value
from (implicit) null to 0 in 3ca8b9968c0 caused the bug
https://github.com/owncloud/music/issues/1251. However, this same principal
issue may have caused also other subtle bugs in various cases not related to
the `dirty` field.