Increase the "batch size" of the online scanning from 10 to 20 tracks
Scanning in larger batches benefits more from the previous optimization
change 5222a63. Furthermore, this and previous optimizations have made the
scanning on MariaDB and PostgreSQL so performant that often the batch of 10
tracks could be handled in around 1 second, of which 300 ms was overhead from
the REST call. Hence, larger batch reduces the total overhead over the entire
library.

Furthermore, also SQLite now performs well enough that batch of 20 songs
shouldn't take so much time that the user would think that the scan has hung
up.