Optimize dirty file checking just under a specified folder
Using `occ music:scan` to rescan dirty files only under a specified path used
to be slow in case there was a large amount of such files. This was because
each dirty file of the library was passed one-by-one to the platform API to
see if it belonged under the specified path. This probably caused at least
one DB query per dirty file.

Now we have custom DB queries to obtain the dirty files only from the
specified folders. Finding recursively all the parent folders to consider is
also well-optimized and should normally require just a handful of DB queries.