Always call closeCursor() after fetching DB query results
According to the documentation
https://docs.nextcloud.com/server/stable/developer_manual/basics/storage/database.html,
this should be done always after execution a custom SELECT query. It also
says that it would not be needed after INSERT/UPDATE/DELETE queries, but on
the other hand, the platform Mapper class does it also there, so presumably
it doesn't hurt either.

The documentation says that failing to close the cursor could cause problems
with SQLite but no such problems have been observed in this project.