Advanced search can search also for radio stations
This is something not included in the advanced search of the Ampache API
specification. However, also our Ampache API implementation actually could
already handle this case and enabling it on the web UI required no major
changes.

On the back-end side, one new advanced search rule "stream_url" was added to
for radio stations.